put API routes in the api namespace
