+++ link-api/config/routes.rb
@@ -1,6 +1,7 @@
 Rails.application.routes.draw do
-  resources :service_at_locations
-  resources :services
-  resources :contacts
-  # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
+  namespace :api do
+    resources :service_at_locations
+    resources :services
+    resources :contacts
+  end
 end
