+++ link-api/spec/routing/service_at_locations_routing_spec.rb
@@ -1,32 +1,32 @@
 require "rails_helper"
 
-RSpec.describe ServiceAtLocationsController, type: :routing do
+RSpec.describe Api::ServiceAtLocationsController, type: :routing do
   describe "routing" do
 
     it "routes to #index" do
-      expect(:get => "/service_at_locations").to route_to("service_at_locations#index")
+      expect(:get => "/api/service_at_locations").to route_to("api/service_at_locations#index")
     end
 
 
     it "routes to #show" do
-      expect(:get => "/service_at_locations/1").to route_to("service_at_locations#show", :id => "1")
+      expect(:get => "/api/service_at_locations/1").to route_to("api/service_at_locations#show", :id => "1")
     end
 
 
     it "routes to #create" do
-      expect(:post => "/service_at_locations").to route_to("service_at_locations#create")
+      expect(:post => "/api/service_at_locations").to route_to("api/service_at_locations#create")
     end
 
     it "routes to #update via PUT" do
-      expect(:put => "/service_at_locations/1").to route_to("service_at_locations#update", :id => "1")
+      expect(:put => "/api/service_at_locations/1").to route_to("api/service_at_locations#update", :id => "1")
     end
 
     it "routes to #update via PATCH" do
-      expect(:patch => "/service_at_locations/1").to route_to("service_at_locations#update", :id => "1")
+      expect(:patch => "/api/service_at_locations/1").to route_to("api/service_at_locations#update", :id => "1")
     end
 
     it "routes to #destroy" do
-      expect(:delete => "/service_at_locations/1").to route_to("service_at_locations#destroy", :id => "1")
+      expect(:delete => "/api/service_at_locations/1").to route_to("api/service_at_locations#destroy", :id => "1")
     end
 
   end
